test_agent/requirements.txt
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langchain>=0.1.0
tavily-python
python-dotenv
pydantic>=1.10.0,<3.0.0